Removed require blog
